third_party/WebKit/LayoutTests/vr/getLayers_notpresenting.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="resources/mock-vr-service.js"></script>
<canvas id="webgl-canvas"></canvas>
<script src="resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  return navigator.getVRDisplays().then( (displays) => {
    var display = displays[0];

    t.step( () => {
      assert_equals(display.getLayers().length, 0);
    }, "No layers initially returned");

    runWithUserGesture( () => {
      display.requestPresent([{ source : webglCanvas }]).then( () => {
        display.exitPresent().then( () => {
          t.step( () => {
            assert_equals(display.getLayers().length, 0);
          }, "No layers returned after exiting");
        }, (err) => {
          t.step( () => {
            assert_unreached();
          }, "exitPresent failed");
        });
      }, (err) => {
        t.step( () => {
          assert_unreached();
        }, "requestPresent failed");
      }).then( () => {
        t.done();
      });
    });
  }, (err) => {
    t.step( () => {
      assert_unreached();
    }, "getVRDisplays rejected");
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"Test getLayers returns an empty array when not presenting");

</script>
